CSS start
I began to add css to the page by first linking the css file to the main index html file so that it could be read and applied. I added an image with a color over it to darken it as it'll be the background for the sign up form.